Add new table to track role provision requests, expand session_user
